emotional-service/requirements.txt
flask
flask-cors
fer
opencv-python-headless
numpy
pillow